refactor: Application에 작성된 로직 LottoGame 클래스로 분리
