+++ README.md
@@ -2,7 +2,7 @@
 
 ## 🚀 기능 목록
 
-### 🎰 Application
+### 🎰 LottoGame
 
 - [x] 로또 번호를 생성한다.
   - [x] 금액을 1000으로 나누어 로또 개수를 구한다.
